requirements.txt
anthropic
openai
python-dotenv
rich
